change mock json data

+++ src/components/Json.js
@@ -11,7 +11,8 @@ const Json = ({ jsonObj, isOwner }) => {
     const ok = window.confirm('Are you sure to delete this json?');
     if (ok) {
       await dbService.doc(`jsons/${jsonObj.id}`).delete();
-      await storageService.refFromURL(jsonObj.attachmentUrl).delete();
+      if (jsonObj.attachmentUrl !== '')
+        await storageService.refFromURL(jsonObj.attachmentUrl).delete();
     }
   };
 
@@ -64,13 +65,14 @@ const Json = ({ jsonObj, isOwner }) => {
         </>
       ) : (
         <>
-          <textarea type="text" rows="15" readOnly>
-            {jsonObj.text}
-          </textarea>
-
-          {jsonObj.attachmentUrl && (
-            <img src={jsonObj.attachmentUrl} alt="attached pic" />
+          {isOwner && (
+            <textarea type="text" rows="15" readOnly value={jsonObj.text} />
           )}
+
+          {/* todo not photo but json */}
+          {/* {jsonObj.attachmentUrl && (
+            <img src={jsonObj.attachmentUrl} alt="attached pic" />
+          )} */}
 
           {isOwner && (
             <div className="json__actions">
